TЗ № 11

+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,63 +1,81 @@
 package ru.yandex.practicum.filmorate.controller;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.service.FilmService;
 
+import javax.validation.Valid;
+import javax.validation.constraints.PositiveOrZero;
 import java.time.LocalDate;
 import java.util.*;
 
+@Validated
 @RestController
 @Slf4j
+@RequiredArgsConstructor
+@RequestMapping("/films")
 public class FilmController {
 
     private final FilmService filmService;
 
-    public FilmController(FilmService filmService) {
-        this.filmService = filmService;
-    }
-
-    @PostMapping("/films")
-    public Film postFilm(@RequestBody Film film) {
+    @PostMapping
+    public Film postFilm(@Valid @RequestBody Film film) {
         validate(film);
         filmService.save(film);
         log.info("Добавлен новый фильм {}", film.getName());
         return film;
     }
 
-    @GetMapping("/films")
-    public List<Film> getFilms() {
+    @GetMapping
+    public Collection<Film> getFilms() {
         return filmService.getAll();
     }
 
-    @GetMapping("/films/{id}")
+    @GetMapping("/{id}")
     public Film getFilmById(@PathVariable Integer id) {
-        return filmService.getById(id);
+            return filmService.getById(id);
     }
 
-    @PutMapping("/films")
-    public Film updateFilm(@RequestBody Film film) {
+    @PutMapping
+    public Film updateFilm(@Valid @RequestBody Film film) {
         validate(film);
-        filmService.update(film);
-        return film;
+        if (filmService.getById(film.getId()) != null) {
+            return filmService.update(film);
+        } else {
+            log.error("Не удалось обновить фильм");
+            throw new NotFoundException("Не удалось обновить. Фильм не найден");
+        }
     }
 
-    @PutMapping("/films/{id}/like/{userId}")
-    public void likeFilm(@PathVariable("id") Integer filmId, @PathVariable("userId") Integer userId) {
-        log.info("Пользователь лайкнул фильм {}", filmService.getById(filmId).getName());
+    @PutMapping("/{id}/like/{userId}")
+    public void likeFilm(@PathVariable("id") @PositiveOrZero Integer filmId,
+                         @PathVariable("userId") @PositiveOrZero Integer userId) {
+        if (filmId < 0 || userId < 0) {
+            log.error("Неверный id");
+            throw new NotFoundException("Неверный id");
+        }
+        log.info("Пользователь лайкнул фильм");
         filmService.addLike(filmId, userId);
     }
 
-    @DeleteMapping("/films/{id}/like/{userId}")
-    public void unlikeFilm(@PathVariable("id") Integer filmId, @PathVariable("userId") Integer userId) {
-        log.info("Пользователь отменил лайк фильму {}", filmService.getById(filmId).getName());
+    @DeleteMapping("/{id}/like/{userId}")
+    public void unlikeFilm(@PathVariable("id") @PositiveOrZero Integer filmId,
+                           @PathVariable("userId") @PositiveOrZero Integer userId) {
+         if (filmId < 0 || userId < 0) {
+            log.error("Неверный id");
+            throw new NotFoundException("Неверный id");
+        }
+        log.info("Пользователь отменил лайк фильму");
         filmService.deleteLike(filmId, userId);
     }
 
-    @GetMapping("/films/popular")
-    public List<Film> getPopularFilms(@RequestParam(defaultValue = "10", required = false) Integer count) {
+    @GetMapping("/popular")
+    public Collection<Film> getPopularFilms(@RequestParam(defaultValue = "10", required = false) Integer count) {
         log.info("Выдан список {} популярных фильмов", count);
         return filmService.getPopular(count);
     }
